Update the plugin to source managed versions from Maven bom
Previously, the plugin used a properties file to source the managed
versions. This commit updates the plugin to replace this with the
use of a dependency management plugin. This allows a Maven bom,
such as the one published by the Platform, to be used. This paves the
way for the removal of the Platform's properties artifact in its 2.0
release.