Ensure that project’s dependency management doesn’t interfere
Previously, the check which verified that every direct dependency was
covered by the Platform used the managed versions from the
springIoTestRuntime configuration. These managed versions included
those configured on it, as well as those inherited from other
configurations’ dependency management and any global dependency
management. If the project being checked also used the dependency
management plugin this could lead to false positives if the project’s
own dependency management included a dependency that wasn’t covered by
the Platform.

This commit upgrades to a new version of the dependency management
plugin that allows a configuration’s own managed versions to be
accessed. Any managed versions declared in global dependency management
or other configurations in the inheritance hierarchy are not included in
these managed versions. This allows the plugin to correctly verify that
a project’s dependencies are included in the Platform.