Recompile tests against Platform deps to avoid binary incompatibilities
Some tests contain code that is source compatible but not binary
compatible with dependency versions provided by the Platform. This
results in test code failing with a NoSuchMethodError at runtime.

While a project's main code must be binary compatible with the
Platform's dependency versions, we don't have to be as stringent
with the test code. This commit updates to plugin to create a new
springIoTest source set that's used to compile code in src/test/java
against the Platform's dependency versions. This allows tests that
are source compatible but not binary compatible with the Platform's
dependency versions to be compiled and run without alteration.