Defer configuration of springIoTest source set until afterEvaluate
Previously, the springIoTest source set was created and configured as
soon as the Spring IO plugin was applied. This meant that any
customisations of the test source set were not reflected in the
springIoTest source set.

This commit defers much of the configuration of the springIoTest
source set until after evaluate so that its configuration can reflect
that of the test source set after any user-applied customisations.

Closes gh-4