Stop disabling dep management plugin's pom customization
Previously, the dependency management plugin's pom customization
would be disabled when this plugin is applied. It's actually
unnecessary as project's using the Spring IO plugin will configure
dependency management for the springIoTestRuntime configuration and
configuration-specific dependency management is not considered
during pom customization.

This commit removes the logic that was disabling the customization.

Closes gh-5